+++ ConcordiumWallet/Views/ScanQR/ScanQRViewController.swift
